Reorganized functions from randomization_utilities.c into several files.
