+++ include/randomization_starters.h
@@ -1,0 +1,8 @@
+#ifndef GUARD_RANDOMIZATION_STARTERS_H
+#define GUARD_RANDOMIZATION_STARTERS_H
+
+u16 GetRandomizedStarterPokemon(u16 chosenStarterId);
+
+u16 GetRivalStarterSpecies(void);
+
+#endif // GUARD_RANDOMIZATION_STARTERS_H
