+++ include/randomization_wild_encounters.h
@@ -1,0 +1,8 @@
+#ifndef GUARD_RANDOMIZATION_WILD_ENCOUNTERS_H
+#define GUARD_RANDOMIZATION_WILD_ENCOUNTERS_H
+
+u8 GetWildMonLevelIncrease(u8 level);
+
+u16 GetRandomizedEncounterSpecies(u16 seedSpecies, u8 level, u8 areaType);
+
+#endif // GUARD_RANDOMIZATION_WILD_ENCOUNTERS_H
