--- include/starter_randomization.h
@@ -1,8 +1,0 @@
-#ifndef GUARD_STARTER_RANDOMIZATION_H
-#define GUARD_STARTER_RANDOMIZATION_H
-
-u16 GetRandomizedStarterPokemon(u16 chosenStarterId);
-
-u16 GetRivalStarterSpecies(void);
-
-#endif // GUARD_STARTER_RANDOMIZATION_H
