+++ src/wild_encounter.c
@@ -5,6 +5,7 @@
 #include "fieldmap.h"
 #include "random.h"
 #include "randomization_utilities.h"
+#include "randomization_wild_encounters.h"
 #include "field_player_avatar.h"
 #include "event_data.h"
 #include "safari_zone.h"
@@ -527,7 +528,7 @@ static bool8 TryGenerateWildMon(const struct WildPokemonInfo *wildMonInfo, u8 ar
 
     // randomization of species here
     species = wildMonInfo->wildPokemon[wildMonIndex].species;
-    species = GetRandomizedSpecies(species, level, area);
+    species = GetRandomizedEncounterSpecies(species, level, area);
     level = GetWildMonLevelIncrease(level);
     species = GetEvolvedWildMonSpecies(species, level);
     CreateWildMon(species, level);
